Move `split_csv_line` to `helpers`
